requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pymupdf>=1.23.0
bertopic>=0.16.0
sentence-transformers>=2.2.0
umap-learn>=0.5.3
hdbscan>=0.8.33
scipy>=1.11.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
scikit-learn>=1.3.0
numpy>=1.24.0
